requirements.txt
boto3
langchain
openai
line-bot-sdk==3.5.0